+++ frontend/src/App.jsx
@@ -1,13 +1,13 @@
 import "./App.scss";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Header from "./components/Header/Header";
-import Footer from "./components/Footer/Footer";
-import HomePage from "./pages/Home/Home";
-import AboutPage from "./pages/About/About";
-import ASLReferencePage from "./pages/ASLReference/ASLReference";
-import PracticePage from "./pages/Practice/Practice";
-import ProfilePage from "./pages/Profile/Profile";
-import NotFoundPage from "./pages/NotFound/NotFound";
+import Header from "./components/Header";
+import Footer from "./components/Footer";
+import HomePage from "./pages/Home";
+import AboutPage from "./pages/About";
+import ASLReferencePage from "./pages/ASLReference";
+import PlayPage from "./pages/Play";
+import ProfilePage from "./pages/Profile";
+import NotFoundPage from "./pages/NotFound";
 
 const Layout = ({ children }) => (
   <div className="screen">
@@ -27,7 +27,7 @@ export default function App() {
           <Route path="/" element={<HomePage />} />
           <Route path="/about" element={<AboutPage />} />
           <Route path="/asl-reference" element={<ASLReferencePage />} />
-          <Route path="/practice" element={<PracticePage />} />
+          <Route path="/play" element={<PlayPage />} />
           {/* Consider using dynamic page names for user profiles. */}
           <Route path="/profile" element={<ProfilePage />} />
           <Route path="*" element={<NotFoundPage />} />
