+++ frontend/src/pages/Play/Play.jsx
@@ -1,20 +1,20 @@
 import React from "react";
-import styles from "./Practice.module.scss";
+import styles from "./Play.module.scss";
 import WebcamSample from "../../components/Camera";
 import Dropdown from "../../components/Dropdown";
 import Timer from "../../components/Timer";
 import Leaderboard from "../../components/Leaderboard";
 
-export default function Practice() {
+export default function Play() {
     return (
         <div>
             <div className={styles.DIDContainer}>
                 <Leaderboard />
                 <div className={styles.introBox}>
-                <h1>Practice Mode</h1>
+                <h1>Play</h1>
                 <p>
                     You will be shown a word and must sign each letter. You will earn more points for
-                    faster signing! There is no time limit and you can practice for as long as you want.
+                    faster signing! There is no time limit and you can play for as long as you want.
                 </p>
                 </div>
                 <Dropdown trigger="Wiki" className={styles.wiki}>
