--- frontend/src/pages/Practice/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Practice";
